# docs/swagger.yaml
basePath: /
definitions:
  domain.Order:
    properties:
      id:
        type: string
      product_id:
        type: string
      quantity:
        type: integer
      total_price:
        type: number
    type: object
  domain.Product:
    properties:
      id:
        type: string
      name:
        type: string
      price:
        type: number
      stock:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a sample server for managing stock and orders.
  title: Inventory & Order Management API
  version: "1.0"
paths:
  /orders:
    get:
      consumes:
      - application/json
      description: Finds all orders
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Order'
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Find all orders
      tags:
      - orders
    post:
      consumes:
      - application/json
      description: Adds a new order to the orders
      parameters:
      - description: Order Info
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/domain.Order'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Order'
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Create a new order
      tags:
      - orders
  /orders/{id}:
    get:
      consumes:
      - application/json
      description: Finds an order by ID
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Order'
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Find an order by ID
      tags:
      - orders
  /products:
    get:
      consumes:
      - application/json
      description: Finds all products
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Product'
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Find all products
      tags:
      - products
    post:
      consumes:
      - application/json
      description: Adds a new product to the inventory
      parameters:
      - description: Product Info
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/domain.Product'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Product'
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Create a new product
      tags:
      - products
  /products/{id}:
    get:
      consumes:
      - application/json
      description: Finds a product by ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Product'
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Find a product by ID
      tags:
      - products
swagger: "2.0"